requirements.txt
discord.py
names
python-dotenv
